Some minor bug fixes.

+++ trunk/src/extrinsics.cpp
@@ -274,7 +274,7 @@ void optimizeCalibrationSets(cv::vector<Size> imSize,
         while (newCorners.at(0).size() < (unsigned int)(num))
         {
 
-            printf("%s << newCorners.at(0).size() = %d; num = %d\n", __FUNCTION__, newCorners.at(0).size(), num);
+            printf("%s << newCorners.at(0).size() = %d; num = %d\n", __FUNCTION__, (int)newCorners.at(0).size(), num);
 
             maxIndex = 0;
             maxScore = 0;
